ml_backend/requirements.txt
Flask==2.3.2
xgboost==1.7.5
pandas==2.0.3
numpy==1.24.2
scikit-learn==1.2.2
joblib==1.2.0
python-binance==1.0.16
